+++ src/main/java/de/alpharogroup/tree/api/ITreeNode.java
@@ -22,7 +22,7 @@
  * OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION
  * WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
  */
-package de.alpharogroup.tree.ifaces;
+package de.alpharogroup.tree.api;
 
 import java.io.Serializable;
 import java.util.ArrayList;