+++ src/test/java/de/alpharogroup/tree/LinkedNodeTest.java
@@ -38,7 +38,7 @@ import org.testng.annotations.BeforeMethod;
 import org.testng.annotations.Test;
 
 import de.alpharogroup.AbstractTestCase;
-import de.alpharogroup.tree.ifaces.ILinkedNode;
+import de.alpharogroup.tree.api.ILinkedNode;
 
 /**
  * The unit test class for the class {@link LinkedNode}